python_api/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv>=1.0.0
tenacity
python-dateutil
tqdm
schedule
lxml
PyMuPDF>=1.23.0
python-docx>=1.1.0
openai>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0